Feat: BirdJudgmentGenerator 구현
